requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.13
openai>=1.10.0
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
chromadb>=0.4.18
tiktoken>=0.5.1
httpx>=0.25.1
pytest>=7.4.3
llama-cpp-python>=0.2.11
sentence-transformers>=2.2.2
tqdm>=4.66.1
llama-stack-client>=0.1.0 